Reorganizing files so that users can see my projects from the main page and don't have to go to /dashboard anymore to see the listing page.
